used constant File.separator instead of "/" or "\\"
