+++ maqs-utilities/src/test/java/com/cognizantsoftvision/maqs/utilities/helper/GenericWaitUnitTest.java
@@ -2,7 +2,7 @@
  * Copyright 2022 (C) Cognizant SoftVision, All rights Reserved
  */
 
-package com.cognizantsoftvision.maqs.utilities;
+package com.cognizantsoftvision.maqs.utilities.helper;
 
 import com.cognizantsoftvision.maqs.utilities.helper.GenericWait;
 import com.cognizantsoftvision.maqs.utilities.helper.TestCategories;
